requirements.txt
fastapi
uvicorn[standard]
bitwarden-sdk
python-dotenv
click
